feat(serializer): Convert file serializer to package.
